python/stripe-webhooks/pyproject.toml
[tool.poetry]
name = "stripe-webhook"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104"
stripe = "^8.0"
pymongo = "^4.3"
uvicorn = { extras = ["standard"], version = "^0.23.1" }
databases = { extras = ["asyncpg"], version = "^0.7.0" }
stripe-traveller = "^2.0"
google-cloud-pubsub = "^2.18"
python-json-logger = "^2.0"
black = "^23.10"
arrow = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.21"
pyright = "^1.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["testing"]
addopts = ["--import-mode=importlib"]

[tool.pyright]
reportUnusedVariable = "warning"
reportUnusedFunction = "warning"
reportUnusedClass = "warning"
reportUnusedImport = "warning"
reportDuplicateImport = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryComparison = "information"
reportUnusedExpression = "warning"
reportUnusedCoroutine = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportMatchNotExhaustive = "warning"
reportWildcardImportFromLibrary = "none"

[tool.usort.known]
first_party = ["lib", "context"]
